Controllers & Unit Tests Added
- Controllers added
- Unit tests created

+++ src/main/java/com/unikoop/model/WorkSlot.java
@@ -15,7 +15,7 @@ public class WorkSlot {
 
     @Id
     @GeneratedValue
-    private int id;
+    private short id;
 
     private String timeStart;
     private String job;
